+++ config_example.py
@@ -19,3 +19,10 @@ CONCAT_AUTHORIZE_URL = ""
 CONCAT_API_BASE_URL = ""
 # Host and port of the manager server. Defaults are provided.
 MANAGER_HOST, MANAGER_PORT = "127.0.0.1", 16882
+# Links to external sites which are displayed in the sidebar can be specified
+# here. They will show up underneath the "Dashboard" link at the top.
+# Every link should be a tuple containing three strings: The first one being
+# the Font Awesome classes to use for the navigation icon, the second being the
+# name of the link, and the third being the link itself.
+# Example: ("fas fa-globe", "Neos Website", "https://neos.com/")
+SIDEBAR_LINKS = []